requirements.txt
fastapi[standard]
requests
pydantic==2.8.2
google-auth==2.33.0
google-generativeai==0.7
bs4==0.0.2
